requirements.txt
opencv-python
Pillow
numpy
qrcode[pil]
reportlab
